+++ changelog.md
@@ -7,6 +7,7 @@
 - Added all bowl foods to `#meals`
 - Added all drinkables to `#drinks`
 - Added Stuffed Cantaloupe to `#feasts`
+- Added all storage blocks to `#forge:storage_blocks`
 - Renamed `Nut Butter Bottle` to `Nut Butter` in English translations
 - Nut Butter now requires 1 Kelp
 - Nut Butter now requires 1 `#forge:salt` if the tag is not empty